requirements.txt
# ===============================================================
# == Зависимости для основного бота (main.py и все коги)      ==
# ===============================================================

# Основная библиотека для работы с Discord API
discord.py>=2.0.0

# Для загрузки переменных из .env файла
python-dotenv

# Для асинхронной работы с базой данных MySQL
aiomysql
PyMySQL # Драйвер, необходимый для aiomysql

# Для работы с API хостинга и вебхуками
requests
aiohttp

# Для парсинга файлов локализации .yml
PyYAML

# Для шифрования токенов
cryptography

# Для получения статистики CPU/RAM
psutil

# telegram
aiogram


# ===============================================================
# == Зависимости для скрипта бэкапов (backup_manager.py)       ==
# ===============================================================

# Официальные библиотеки Google для работы с Google Drive API
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

# Для работы с Redis
redis[asyncio]

humanize

aiofiles

prometheus-client

# Для создания веб-API
fastapi
uvicorn[standard]